+++ solidity/src/interfaces/PowersEvents.sol
@@ -85,12 +85,12 @@ interface PowersEvents {
     /// @param roleId the id of the role
     /// @param account the address of the account that has the role
     /// @param access whether the account has access to the role
-    event RoleSet(uint32 indexed roleId, address indexed account, bool indexed access);
+    event RoleSet(uint256 indexed roleId, address indexed account, bool indexed access);
 
     /// @notice Emitted when a role is labelled.
     /// @param roleId the id of the role
     /// @param label the label assigned to the role
-    event RoleLabel(uint32 indexed roleId, string label);
+    event RoleLabel(uint256 indexed roleId, string label);
 
     /// @notice Emitted when a law is adopted.
     /// @param law the address of the law
