+++ solidity/src/laws/bespoke/governYourTax/Grant.sol
@@ -37,7 +37,7 @@ contract Grant is Law {
         string memory name_,
         string memory description_,
         address payable powers_,
-        uint32 allowedRole_,
+        uint256 allowedRole_,
         LawChecks memory config_,
         //
         uint48 duration_,
