+++ solidity/src/laws/executive/BespokeAction.sol
@@ -44,7 +44,7 @@ contract BespokeAction is Law {
         string memory name_,
         string memory description_,
         address payable powers_,
-        uint32 allowedRole_,
+        uint256 allowedRole_,
         LawChecks memory config_,
         // specific parameters
         address targetContract_,
