+++ solidity/src/laws/executive/PresetAction.sol
@@ -44,7 +44,7 @@ contract PresetAction is Law {
         string memory name_,
         string memory description_,
         address payable powers_,
-        uint32 allowedRole_,
+        uint256 allowedRole_,
         LawChecks memory config_,
         // specific to preset action
         address[] memory targets_,
