+++ solidity/src/laws/executive/SelfDestructAction.sol
@@ -30,7 +30,7 @@ contract SelfDestructAction is Law {
         string memory name_,
         string memory description_,
         address payable powers_,
-        uint32 allowedRole_,
+        uint256 allowedRole_,
         LawChecks memory config_, 
         address[] memory targets_,
         uint256[] memory values_,
